Working tooltips with ember-popper (popper.js)
