+++ addon/components/ember-popover.js
@@ -1,0 +1,4 @@
+import PopperBaseComponent from 'ember-tooltips/components/popper-base';
+
+export default PopperBaseComponent.extend({
+});
