--- addon/components/popover-on-component.js
@@ -1,7 +1,0 @@
-import PopoverOnElementComponent from 'ember-tooltips/components/popover-on-element';
-
-export default PopoverOnElementComponent.extend({
-  tetherComponentName: 'tether-popover-on-component',
-
-  _shouldTargetGrandparentView: true,
-});
