--- addon/components/popover-on-element.js
@@ -1,5 +1,0 @@
-import LazyRenderWrapperComponent from 'ember-tooltips/components/lazy-render-wrapper';
-
-export default LazyRenderWrapperComponent.extend({
-  tetherComponentName: 'tether-popover-on-element',
-});
