--- addon/components/tether-popover-on-component.js
@@ -1,6 +1,0 @@
-import TetherPopoverOnElement from 'ember-tooltips/components/tether-popover-on-element';
-import { onComponentTarget } from 'ember-tooltips/utils';
-
-export default TetherPopoverOnElement.extend({
-  target: onComponentTarget,
-});
