--- addon/components/tether-tooltip-on-component.js
@@ -1,6 +1,0 @@
-import TooltipOnElementComponent from 'ember-tooltips/components/tether-tooltip-on-element';
-import { onComponentTarget } from 'ember-tooltips/utils';
-
-export default TooltipOnElementComponent.extend({
-  target: onComponentTarget,
-});
