+++ addon/styles/addon.scss
@@ -34,70 +34,70 @@
   white-space: normal; // IE doesnt support initial so fall back to normal
   white-space: initial;
 
-  &:after {
-    content: '';
-    position: absolute;
-    width: 10px;
-    height: 10px;
-    margin: -5px;
-    background: inherit;
-    -webkit-transform: rotate(45deg);
-    -ms-transform: rotate(45deg);
-    transform: rotate(45deg);
-  }
+  // &:after {
+  //   content: '';
+  //   position: absolute;
+  //   width: 10px;
+  //   height: 10px;
+  //   margin: -5px;
+  //   background: inherit;
+  //   -webkit-transform: rotate(45deg);
+  //   -ms-transform: rotate(45deg);
+  //   transform: rotate(45deg);
+  // }
 
-  &.ember-tooltip-or-popover-target-attached-top:after,
-  &.ember-tooltip-or-popover-target-attached-top-left:after,
-  &.ember-tooltip-or-popover-target-attached-top-right:after {
-    bottom: 0;
-  }
-  &.ember-tooltip-or-popover-target-attached-bottom:after,
-  &.ember-tooltip-or-popover-target-attached-bottom-left:after,
-  &.ember-tooltip-or-popover-target-attached-bottom-right:after {
-    top: 0;
-  }
-  &.ember-tooltip-or-popover-target-attached-top:after,
-  &.ember-tooltip-or-popover-target-attached-bottom:after {
-    left: 50%;
-  }
-  &.ember-tooltip-or-popover-target-attached-top-left:after,
-  &.ember-tooltip-or-popover-target-attached-bottom-left:after {
-    right: 15px;
-  }
-  &.ember-tooltip-or-popover-target-attached-top-right:after,
-  &.ember-tooltip-or-popover-target-attached-bottom-right:after {
-    left: 15px;
-  }
+  // &.ember-tooltip-or-popover-target-attached-top:after,
+  // &.ember-tooltip-or-popover-target-attached-top-left:after,
+  // &.ember-tooltip-or-popover-target-attached-top-right:after {
+  //   bottom: 0;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-bottom:after,
+  // &.ember-tooltip-or-popover-target-attached-bottom-left:after,
+  // &.ember-tooltip-or-popover-target-attached-bottom-right:after {
+  //   top: 0;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-top:after,
+  // &.ember-tooltip-or-popover-target-attached-bottom:after {
+  //   left: 50%;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-top-left:after,
+  // &.ember-tooltip-or-popover-target-attached-bottom-left:after {
+  //   right: 15px;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-top-right:after,
+  // &.ember-tooltip-or-popover-target-attached-bottom-right:after {
+  //   left: 15px;
+  // }
 
-  &.ember-tooltip-or-popover-target-attached-left:after,
-  &.ember-tooltip-or-popover-target-attached-left-top:after,
-  &.ember-tooltip-or-popover-target-attached-left-bottom:after {
-    right: 0;
-  }
-  &.ember-tooltip-or-popover-target-attached-right:after,
-  &.ember-tooltip-or-popover-target-attached-right-top:after,
-  &.ember-tooltip-or-popover-target-attached-right-bottom:after {
-    left: 0;
-  }
-  &.ember-tooltip-or-popover-target-attached-left:after,
-  &.ember-tooltip-or-popover-target-attached-right:after {
-    top: 50%;
-  }
-  &.ember-tooltip-or-popover-target-attached-left-top:after,
-  &.ember-tooltip-or-popover-target-attached-right-top:after {
-    bottom: 15px;
-  }
-  &.ember-tooltip-or-popover-target-attached-left-bottom:after,
-  &.ember-tooltip-or-popover-target-attached-right-bottom:after {
-    top: 15px;
-  }
+  // &.ember-tooltip-or-popover-target-attached-left:after,
+  // &.ember-tooltip-or-popover-target-attached-left-top:after,
+  // &.ember-tooltip-or-popover-target-attached-left-bottom:after {
+  //   right: 0;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-right:after,
+  // &.ember-tooltip-or-popover-target-attached-right-top:after,
+  // &.ember-tooltip-or-popover-target-attached-right-bottom:after {
+  //   left: 0;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-left:after,
+  // &.ember-tooltip-or-popover-target-attached-right:after {
+  //   top: 50%;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-left-top:after,
+  // &.ember-tooltip-or-popover-target-attached-right-top:after {
+  //   bottom: 15px;
+  // }
+  // &.ember-tooltip-or-popover-target-attached-left-bottom:after,
+  // &.ember-tooltip-or-popover-target-attached-right-bottom:after {
+  //   top: 15px;
+  // }
 }
 
 /* Fade */
 
 .ember-tooltip,
 .ember-popover {
-  opacity: 0;
+  // opacity: 0;
 
   &.ember-tooltip-or-popover-slide,
   &.ember-tooltip-or-popover-fade {
