--- addon/templates/components/lazy-render-wrapper.hbs
@@ -1,20 +1,0 @@
-{{#if _shouldRender}}
-	{{#component tetherComponentName
-		passedPropertiesObject=passedPropertiesObject
-		class=class
-		id=id
-		classNames=classNames
-
-		_shouldShowOnRender=_shouldShowOnRender
-		_isInProcessOfShowing=_isInProcessOfShowing
-	}}
-    {{#if hasBlock}}
-      {{yield (hash
-        hide=(action "hide")
-      )}}
-    {{else}}
-      {{text}}
-    {{/if}}
-	{{/component}}
-{{/if}}
-
