--- app/components/base-component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-tooltips/components/base-component';
