--- app/components/popover-on-element.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-tooltips/components/popover-on-element';
