--- app/components/tether-popover-on-element.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-tooltips/components/tether-popover-on-element';
