--- app/components/tooltip-on-component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-tooltips/components/tooltip-on-component';
