+++ tests/dummy/app/styles/app.scss
@@ -10,6 +10,7 @@ html {
 
 body {
   font-size: #{$bodyFontSize}rem;
+  background-color: #eee; /* TODO - remove */
 
   /* So text looks much better in Crome/Safari */
